restructured  api/activities, api/operators and api/graphs
