+++ frog/imports/ui/GraphEditor/SidePanel/OperatorPanel/ChooseOperator.jsx
@@ -10,7 +10,7 @@ import { withStyles } from '@material-ui/core/styles';
 import Search from '@material-ui/icons/Search';
 
 import type { operatorPackageT, OperatorDbT } from 'frog-utils';
-import { Operators } from '/imports/api/activities';
+import { Operators } from '/imports/api/operators';
 import { operatorTypes, operatorTypesObj } from '/imports/operatorTypes';
 import { type StoreProp } from '../../store';
 import ListComponent from '../ListComponent';
