+++ frog/imports/ui/GraphEditor/SidePanel/OperatorPanel/index.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { withTracker } from 'meteor/react-meteor-data';
 
-import { Operators } from '/imports/api/activities';
+import { Operators } from '/imports/api/operators';
 import { connect } from '../../store';
 
 import EditClass from './EditOperator';
