+++ frog/imports/ui/TeacherView/utils/exportComponent.js
@@ -7,12 +7,13 @@ import FileSaver from 'file-saver';
 import { cleanEmptyCols, type ActivityDbT, strfTime } from 'frog-utils';
 import { omit } from 'lodash';
 
-import { Sessions } from '../../../api/sessions';
-import { Objects } from '../../../api/objects';
-import { Activities, Operators } from '../../../api/activities';
-import { Products } from '../../../api/products';
-import { activityTypesObj } from '../../../activityTypes';
-import { getActivitySequence } from '../../../api/graphSequence';
+import { Sessions } from '/imports/api/sessions';
+import { Objects } from '/imports/api/objects';
+import { Activities } from '/imports/api/activities';
+import { Operators } from '/imports/api/operators';
+import { Products } from '/imports/api/products';
+import { activityTypesObj } from '/imports/activityTypes';
+import { getActivitySequence } from '/imports/api/graphSequence';
 import { graphToString } from '../../GraphEditor/utils/export';
 import downloadLog from './downloadLog';
 import exportGraphPNG from '../../GraphEditor/utils/exportPicture';
