+++ frog/server/externalOperators.js
@@ -2,7 +2,7 @@
 import redis from 'redis';
 import { Meteor } from 'meteor/meteor';
 import { operatorTypes, operatorTypesObj } from '../imports/operatorTypes';
-import { ExternalOperators } from '../imports/api/activities';
+import { ExternalOperators } from '../imports/api/operators';
 
 export default () => {
   ExternalOperators.find({})
