+++ frog/server/runDataflow.js
@@ -14,7 +14,8 @@ import mergeData from './mergeData';
 import reactiveToProduct from './reactiveToProduct';
 import { operatorTypesObj, operators } from '../imports/operatorTypes';
 import { Products } from '../imports/api/products';
-import { Operators, Activities, Connections } from '../imports/api/activities';
+import { Activities, Connections } from '../imports/api/activities';
+import { Operators } from '../imports/api/operators';
 import { addObject } from '../imports/api/objects';
 import remote from './runRemoteOperator';
